# String.Replace optimisation