--- a/NExtends/NExtends.csproj
+++ b/NExtends/NExtends.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
+    <TargetFrameworks>netstandard2.0;netcoreapp2.0;net461</TargetFrameworks>
     <VersionPrefix>2.0.0</VersionPrefix>
     <VersionSuffix>alpha</VersionSuffix>
     <PackageVersion>2.0.0</PackageVersion>