2.07x optimisation for tested case
Resolve https://github.com/LuccaSA/NExtends/issues/29

In netcore 2, string replace with StringComparison already exists, so we don't provide it for NETCOREAPP2 target.
https://github.com/dotnet/coreclr/blob/master/src/mscorlib/src/System/String.Manipulation.cs#L932

For net461 target, here are the result of a litlle benchmark : 
``` ini

BenchmarkDotNet=v0.10.10, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.64)
Processor=Intel Core i7-6700HQ CPU 2.60GHz (Skylake), ProcessorCount=8
Frequency=2531250 Hz, Resolution=395.0617 ns, Timer=TSC
.NET Core SDK=2.0.3
  [Host] : .NET Core 2.0.3 (Framework 4.6.25815.02), 64bit RyuJIT  [AttachedDebugger]
  Clr    : .NET Framework 4.6.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2600.0
  Core   : .NET Core 2.0.3 (Framework 4.6.25815.02), 64bit RyuJIT


```
|          Method |  Job | Runtime |       Mean |     Error |    StdDev | Scaled | ScaledSD |  Gen 0 | Allocated |
|---------------- |----- |-------- |-----------:|----------:|----------:|-------:|---------:|-------:|----------:|
|          Dotnet (not StringComparison) |  Clr |     Clr |   274.1 ns |  5.449 ns |  9.686 ns |   1.00 |     0.00 | 0.0582 |     184 B |
|        NExtends |  Clr |     Clr | 2,304.8 ns | 38.344 ns | 35.867 ns |   8.42 |     0.31 | 0.5951 |    1880 B |
|         Recurse |  Clr |     Clr | 1,651.1 ns | 24.343 ns | 21.580 ns |   6.03 |     0.22 | 0.4387 |    1384 B |
|  RecurseBuilder |  Clr |     Clr | 1,260.6 ns | 17.872 ns | 16.717 ns |   4.60 |     0.17 | 0.2308 |     728 B |
| RecurseBuilder2 |  Clr |     Clr | 1,109.3 ns |  5.630 ns |  4.071 ns |   4.05 |     0.14 | 0.1316 |     416 B |
|          Dotnet | Core |    Core |   921.4 ns | 18.373 ns | 23.236 ns |   1.00 |     0.00 | 0.0582 |     184 B |
|        NExtends | Core |    Core | 2,150.6 ns | 30.726 ns | 28.741 ns |   2.34 |     0.06 | 0.3548 |    1120 B |
|         Recurse | Core |    Core | 1,195.2 ns | 28.481 ns | 40.847 ns |   1.30 |     0.05 | 0.4387 |    1384 B |
|  RecurseBuilder | Core |    Core | 1,154.0 ns | 15.152 ns | 13.432 ns |   1.25 |     0.03 | 0.2308 |     728 B |
| RecurseBuilder2 | Core |    Core | 1,021.2 ns | 14.570 ns | 12.916 ns |   1.11 |     0.03 | 0.1316 |     416 B |

Here the test methods : 

```
    public static class StringExtensions
    {
        public static string ReplaceDotNet(this string originalString, string oldValue, string newValue, StringComparison comparisonType)
        {
#if NETCOREAPP2_0
            return originalString.Replace(oldValue, newValue, comparisonType);
#else
            return originalString.Replace(oldValue, newValue);
#endif
        }

        public static string ReplaceRegex(this string originalString, string oldValue, string newValue, StringComparison comparisonType)
        {
            var regexOptions = RegexOptions.IgnoreCase;

            if (comparisonType == StringComparison.CurrentCulture
                || comparisonType == StringComparison.InvariantCulture
                || comparisonType == StringComparison.Ordinal)
            {
                regexOptions = RegexOptions.CultureInvariant;
            }

            return Regex.Replace(
                originalString,
                Regex.Escape(oldValue),
                newValue.Replace("$", "$$"),
                regexOptions
            );
        }

        public static string ReplaceRecurse(this string source, string oldString, string newString, StringComparison comparison, bool recursive = true)
        {
            int index = source.IndexOf(oldString, comparison);

            while (index > -1)
            {
                source = source.Remove(index, oldString.Length);
                source = source.Insert(index, newString);

                if (!recursive)
                {
                    return source;
                }
                index = source.IndexOf(oldString, index + newString.Length, comparison);
            }

            return source;
        }

        public static string ReplaceRecurseStringBuilder(this string source, string oldValue, string newValue, StringComparison comparisonType)
        {
            if (source.Length == 0 || oldValue.Length == 0)
                return source;

            var result = new StringBuilder();
            int startingPos = 0;
            int nextMatch;
            while ((nextMatch = source.IndexOf(oldValue, startingPos, comparisonType)) > -1)
            {
                result.Append(source, startingPos, nextMatch - startingPos);
                result.Append(newValue);
                startingPos = nextMatch + oldValue.Length;
            }
            result.Append(source, startingPos, source.Length - startingPos);

            return result.ToString();
        }

        public static string ReplaceRecurseStringBuilder2(this string str, string oldValue, string newValue, StringComparison comparison)
        {
            if (oldValue == null)
                throw new ArgumentNullException("oldValue");
            if (oldValue.Length == 0)
                throw new ArgumentException("String cannot be of zero length.", "oldValue");

            StringBuilder sb = null;

            int startIndex = 0;
            int foundIndex = str.IndexOf(oldValue, comparison);
            while (foundIndex != -1)
            {
                if (sb == null)
                    sb = new StringBuilder(str.Length + (newValue != null ? Math.Max(0, 5 * (newValue.Length - oldValue.Length)) : 0));
                sb.Append(str, startIndex, foundIndex - startIndex);
                sb.Append(newValue);

                startIndex = foundIndex + oldValue.Length;
                foundIndex = str.IndexOf(oldValue, startIndex, comparison);
            }

            if (startIndex == 0)
                return str;
            sb.Append(str, startIndex, str.Length - startIndex);
            return sb.ToString();
        }

    }
```

Test call : 

```
   [CoreJob, ClrJob]
    [MarkdownExporter, MemoryDiagnoser]
    public class BenchmarkStringReplace
    {
        private string data = "This is a not very very long string, but have very very nice readability";
        private string find = "very";
        private string replace = "super";

        [Benchmark(Baseline = true)]
        public void Dotnet()
        {
            data.ReplaceDotNet(find, replace, StringComparison.CurrentCultureIgnoreCase);
        }
        [Benchmark]
        public void NExtends()
        {
            data.ReplaceRegex(find, replace, StringComparison.CurrentCultureIgnoreCase);
        }
        [Benchmark]
        public void Recurse()
        {
            data.ReplaceRecurse(find, replace, StringComparison.CurrentCultureIgnoreCase);
        }
        [Benchmark]
        public void RecurseBuilder()
        { 
            data.ReplaceRecurseStringBuilder(find, replace, StringComparison.CurrentCultureIgnoreCase);
        }
        [Benchmark]
        public void RecurseBuilder2()
        {
            data.ReplaceRecurseStringBuilder2(find, replace, StringComparison.CurrentCultureIgnoreCase);
        }
    }
```